--- core/shims/spark321/src/main/resources/META-INF/services/com.intel.raydp.shims.SparkShimProvider
@@ -1,1 +1,0 @@
-com.intel.raydp.shims.spark321.SparkShimProvider
