+++ core/shims/spark322/src/main/resources/META-INF/services/com.intel.raydp.shims.SparkShimProvider
@@ -1,0 +1,1 @@
+com.intel.raydp.shims.spark322.SparkShimProvider
